requirements.txt
pandas==1.5.3
numpy==1.23.5
wordcloud==1.9.2
matplotlib==3.7.0
numerize==0.12
Pillow==9.4.0
pyarrow==11.0.0
streamlit==1.27.2
streamlit-option-menu==0.3.6
plotly==5.18.0